Revert "Created Keycloak client"
This reverts commit 4ed265f2a68b5f9869a6731e6ec5e4139c26ccef.
