+++ pom.xml
@@ -87,18 +87,6 @@
       <artifactId>caffeine</artifactId>
       <version>2.6.2</version>
     </dependency>
-    <dependency>
-	    <groupId>org.powermock</groupId>
-	    <artifactId>powermock-module-junit4</artifactId>
-	    <version>1.6.4</version>
-	    <scope>test</scope>
-	  </dependency>
-	  <dependency>
-	    <groupId>org.powermock</groupId>
-	    <artifactId>powermock-api-mockito</artifactId>
-	    <version>1.6.4</version>
-	    <scope>test</scope>
-	  </dependency>
   </dependencies>
 
   <build>
