+++ src/main/java/com/cpqd/app/auth/Auth.java
@@ -1,17 +1,17 @@
 package com.cpqd.app.auth;
 
-import java.util.ArrayList;
-
-import org.json.JSONArray;
-import org.json.JSONException;
-
-import com.cpqd.app.config.Config;
 import com.mashape.unirest.http.HttpResponse;
 import com.mashape.unirest.http.JsonNode;
 import com.mashape.unirest.http.Unirest;
 import com.mashape.unirest.http.exceptions.UnirestException;
+import com.cpqd.app.config.Config;
+import org.apache.commons.codec.binary.Base64;
+import org.json.JSONArray;
+import org.json.JSONException;
 
-public class Auth extends TenantsOwner {
+import java.util.ArrayList;
+
+public class Auth {
 
     private static Auth mInstance;
 
@@ -24,7 +24,33 @@ public class Auth extends TenantsOwner {
         return mInstance;
     }
 
-    @Override
+    /**
+     * Generate token to be used in internal communication.
+     *
+     * @param tenant is the dojot tenant for which the token should be valid for
+     * @return JWT token to be used in the requests
+     */
+    public String getToken(String tenant) {
+        StringBuffer payload = new StringBuffer("{\"service\":\"");
+        payload.append(tenant);
+        payload.append("\",\"username\":\"iotagent\"}");
+
+        System.out.println(payload);
+
+        StringBuffer response = new StringBuffer(Base64.encodeBase64String("jwt schema".getBytes()));
+        response.append(".");
+        response.append(Base64.encodeBase64String(payload.toString().getBytes()));
+        response.append(".");
+        response.append(Base64.encodeBase64String("dummy signature".getBytes()));
+
+        return response.toString();
+    }
+
+    /**
+     * Gets all tenants that are registered on Auth service.
+     *
+     * @return List of tenants.
+     */
     public ArrayList<String> getTenants(){
         StringBuffer url = new StringBuffer(Config.getInstance().getAuthAddress());
         url.append("/admin/tenants");
