+++ src/main/java/com/cpqd/app/config/Config.java
@@ -21,12 +21,6 @@ public class Config {
     private Integer mKafkaDefaultSessionTimeout;
     private String mKafkaDefaultGroupId;
     private Long mKafkaDefaultConsumerPollTime;
-    
-    private String keycloakBasePath;
-    private String keycloakUsername;
-    private String keycloakPassword;
-    private String keycloakClientId;
-    private String keycloakGrantType;
 
     private Config() {
         if (System.getenv("DEVM_ADDRESS") != null) {
@@ -63,36 +57,6 @@ public class Config {
             this.mKafkaDefaultConsumerPollTime = new Long(System.getenv("KAFKA_CONSUMER_POLL_TIME"));
         } else {
             this.mKafkaDefaultConsumerPollTime = new Long(100);
-        }
-        
-        if (System.getenv("KEYCLOAK_BASE_PATH") != null) {
-        	this.keycloakBasePath = System.getenv("KAFKA_CONSUMER_POLL_TIME");
-        }else {
-        	this.keycloakBasePath = "http://localhost:8080/auth";
-        }
-        
-        if (System.getenv("KEYCLOAK_USERNAME") != null) {
-        	this.keycloakUsername = System.getenv("KEYCLOAK_USERNAME");
-        }else {
-        	this.keycloakUsername = "admin";
-        }
-        
-        if (System.getenv("KEYCLOAK_PASSWORD") != null) {
-        	this.keycloakPassword = System.getenv("KEYCLOAK_PASSWORD");
-        }else {
-        	this.keycloakPassword = "admin";
-        }
-        
-        if (System.getenv("KEYCLOAK_CLIENT_ID") != null) {
-        	this.keycloakClientId = System.getenv("KEYCLOAK_CLIENT_ID");
-        }else {
-        	this.keycloakClientId = "admin-cli";
-        }
-        
-        if (System.getenv("KEYCLOAK_GRANT_TYPE") != null) {
-        	this.keycloakGrantType = System.getenv("KEYCLOAK_GRANT_TYPE");
-        }else {
-        	this.keycloakGrantType = "password";
         }
 
 
@@ -179,45 +143,4 @@ public class Config {
     public long getKafkaDefaultConsumerPollTime() {
         return this.mKafkaDefaultConsumerPollTime.longValue();
     }
-
-	public String getKeycloakBasePath() {
-		return keycloakBasePath;
-	}
-
-	public void setKeycloakBasePath(String keycloakBasePath) {
-		this.keycloakBasePath = keycloakBasePath;
-	}
-
-	public String getKeycloakUsername() {
-		return keycloakUsername;
-	}
-
-	public void setKeycloakUsername(String keycloakUsername) {
-		this.keycloakUsername = keycloakUsername;
-	}
-
-	public String getKeycloakPassword() {
-		return keycloakPassword;
-	}
-
-	public void setKeycloakPassword(String keycloakPassword) {
-		this.keycloakPassword = keycloakPassword;
-	}
-
-	public String getKeycloakClientId() {
-		return keycloakClientId;
-	}
-
-	public void setKeycloakClientId(String keycloakClientId) {
-		this.keycloakClientId = keycloakClientId;
-	}
-
-	public String getKeycloakGrantType() {
-		return keycloakGrantType;
-	}
-
-	public void setKeycloakGrantType(String keycloakGrantType) {
-		this.keycloakGrantType = keycloakGrantType;
-	}
-    
 }
